ml-service/app/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.32.0
sentence-transformers==3.1.1
pandas==2.2.3
scikit-learn==1.5.2
numpy==2.1.3
torch==2.5.0
psycopg2-binary==2.9.10
shap==0.48.0
langdetect==1.0.9
python-multipart==0.0.9
umap-learn==0.5.6
scipy==1.14.1
sqlalchemy==2.0.36
transformers==4.45.2
pydantic==2.9.2
accelerate>=0.26.0
